Properly wait for WiFi
Corrected the problem that the arduino would not recover from a power failure.  The WiFi router would take longer than the arduino to reboot and the arduino code didn't properly wait for the WiFi connection to return.
I also added code to reset the arduino if the data isn't updated.  Likely this is because we lost WiFi, but not power.